fbdev: msm: mdss: Implement link clocks reconfiguration hack
This hack sets byte and pixel clocks to half rate and back
to normal rate when property qcom,link-clks-reconf-platform-hack
is found in the mdss-dsi-ctrl node.
It is required (temporarily) for MSM8998 on the common clock
framework due to an unknown bug appearing on suspend/resume
producing inability to resume the display.