clk: qcom: mmcc-msm8998: Enable safe config for video_core_clk_src
When watching youtube shorts, video playback often times glitch as the clk
tend to struggle turning on and off while scrolling as it sometimes fail to
re-enable for the new instance; this causes the playback to instead get stuck
with the previous instance rather than loading the new one. This fixes the
issue as it makes the clk park at XO and reconfigure the parent clk properly
before re-enabling. Remove FORCE_ENABLE_RCG as it conflicts with safe config.

Signed-off-by: ederekun <sedrickvince@gmail.com>