clk: gpucc-msm8998: Repair vdd_gpucc regulator corner scaling
Based on k4.4 DT, various tests (undoubling of frequencies, reducing fmax
definitions, and intentionally mismatching fmax definition to frequencies),
and analysis of the GPU frequency table's relation to the regulator,
a non-linear progression of the GPU voltage has been observed to be needed
by the GPU. This is due to the doubling of frequencies which is necessary
for achieving the full performance of the GPU, unfortunately the voltage
scaling and kgsl are configured for the non-doubled frequencies. For
this reason, we need to reach 8 uV around SVS level to initialize kgsl
at the right OPP min and max; let's scale the 4 levels below 710MHz according
to the qcom,cpr-corner-fmax-map and have the regulator run at max uV for
frequencies beyond 710MHz.

Signed-off-by: ederekun <sedrickvince@gmail.com>