clk: qcom: cpu-osm-legacy: Set CLK_SET_RATE_PARENT for aux clk
This finally fixes the issue by which the OSM driver is not
setting the hmss_gpll0 clk to 300MHz, leading to developers
relying on setting it at GCC clock probe which isn't right.
Simply define CLK_SET_RATE_PARENT for the aux clk to have the
clock framework propagate the init rate to the hmss_gpll0 clk
properly as intended.

Signed-off-by: ederekun <sedrickvince@gmail.com>