clk: gcc/mmcc-msm8998: Complete alpha clks' configuration
We have missing configurations that seem to have been missed;
define them for the ops to have proper values when managing
the clks. Also switch the post divs for mmcc clks to the header's
fabia post div as it is more complete and is the one used by the
gcc gpll out_main clks.

Signed-off-by: ederekun <sedrickvince@gmail.com>