arm64: DT: msm8998: Fix BCL sensors configuration
k4.4 only has vbat and ibat configured, this is reflected by the
reconfiguration done after importing other msm8998 DTs. However,
for some reason, we have 5 BCL clients configured in pmi8998 which
are not likely to work properly given the post-reconfiguration;
the 0x0 and 0x2 BCL clients are the only ones that are valid based
on the k4.4 reconfiguration.

In addition to this, Oneplus/Sony thermals configuration tries to
configure the 'soc' subnode under thermals using bcl_sensor 0x2.
Unfortunately, this is not valid as the 'temperature' values 0x2
gives do not reflect what is assumed to be battery percentage but
rather battery voltage, making the trip values nonsensical.

Signed-off-by: ederekun <sedrickvince@gmail.com>