arm64: DT: pmi8998: Remove C1 shut down on low_vbat trip
With reference to Oneplus/Sony thermal commons, these entries instruct
the thermal driver to shut down C1 when the battery voltage trips. This
behavior is not seen with any other SoC for vbat and I am not certain
why this was added. Let's follow QCOM's sdm660/pm660 configuration and just
setup the trip for vbat.

Signed-off-by: ederekun <sedrickvince@gmail.com>